Side navigation view is working
Toggle is working, planning on having back button functionality by overwriting the actionbar when needed instead of embedding that in the actionbar itself. The custom action bar will then be added back in when necessary. That's the plan at least.

Currently anything other than home crashes the app, need to add locations to the listener